+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/KryptonAboutToolkitForm.Designer.cs
@@ -60,6 +60,8 @@
             this.tsbtnDeveloperInformation = new System.Windows.Forms.ToolStripButton();
             this.tssVersions = new System.Windows.Forms.ToolStripSeparator();
             this.tsbtnVersions = new System.Windows.Forms.ToolStripButton();
+            this.clmnFileName = new System.Windows.Forms.DataGridViewTextBoxColumn();
+            this.clmnVersion = new System.Windows.Forms.DataGridViewTextBoxColumn();
             ((System.ComponentModel.ISupportInitialize)(this.kpnlButtons)).BeginInit();
             this.kpnlButtons.SuspendLayout();
             this.tableLayoutPanel1.SuspendLayout();
@@ -388,6 +390,9 @@
             this.kdgvVersions.AutoSizeColumnsMode = System.Windows.Forms.DataGridViewAutoSizeColumnsMode.AllCells;
             this.kdgvVersions.AutoSizeRowsMode = System.Windows.Forms.DataGridViewAutoSizeRowsMode.AllCells;
             this.kdgvVersions.BorderStyle = System.Windows.Forms.BorderStyle.None;
+            this.kdgvVersions.Columns.AddRange(new System.Windows.Forms.DataGridViewColumn[] {
+            this.clmnFileName,
+            this.clmnVersion});
             this.kdgvVersions.Dock = System.Windows.Forms.DockStyle.Fill;
             this.kdgvVersions.Location = new System.Drawing.Point(0, 0);
             this.kdgvVersions.Name = "kdgvVersions";
@@ -409,7 +414,7 @@
             this.tsbtnVersions});
             this.tsControls.Location = new System.Drawing.Point(3, 0);
             this.tsControls.Name = "tsControls";
-            this.tsControls.Size = new System.Drawing.Size(465, 25);
+            this.tsControls.Size = new System.Drawing.Size(434, 25);
             this.tsControls.TabIndex = 0;
             // 
             // tsbtnGeneralInformation
@@ -470,6 +475,18 @@
             this.tsbtnVersions.Text = "Version Information";
             this.tsbtnVersions.Click += new System.EventHandler(this.tsbtnVersions_Click);
             // 
+            // clmnFileName
+            // 
+            this.clmnFileName.HeaderText = "Column1";
+            this.clmnFileName.Name = "clmnFileName";
+            this.clmnFileName.Width = 85;
+            // 
+            // clmnVersion
+            // 
+            this.clmnVersion.HeaderText = "Column1";
+            this.clmnVersion.Name = "clmnVersion";
+            this.clmnVersion.Width = 85;
+            // 
             // KryptonAboutToolkitForm
             // 
             this.AcceptButton = this.kbtnOk;
@@ -479,11 +496,16 @@
             this.Controls.Add(this.kryptonPanel1);
             this.Controls.Add(this.kpnlButtons);
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.Fixed3D;
+            this.FormTitleAlign = Krypton.Toolkit.PaletteRelativeAlign.Near;
+            this.GroupBackStyle = Krypton.Toolkit.PaletteBackStyle.FormMain;
+            this.GroupBorderStyle = Krypton.Toolkit.PaletteBorderStyle.FormMain;
+            this.HeaderStyle = Krypton.Toolkit.HeaderStyle.Form;
             this.MaximizeBox = false;
             this.MinimizeBox = false;
             this.Name = "KryptonAboutToolkitForm";
             this.ShowIcon = false;
             this.ShowInTaskbar = false;
+            this.TitleStyle = Krypton.Toolkit.KryptonFormTitleStyle.Inherit;
             ((System.ComponentModel.ISupportInitialize)(this.kpnlButtons)).EndInit();
             this.kpnlButtons.ResumeLayout(false);
             this.kpnlButtons.PerformLayout();
@@ -556,5 +578,7 @@
         private ToolStripButton tsbtnDeveloperInformation;
         private ToolStripSeparator tssVersions;
         private ToolStripButton tsbtnVersions;
+        private DataGridViewTextBoxColumn clmnFileName;
+        private DataGridViewTextBoxColumn clmnVersion;
     }
 }